+++ app/.husky/pre-commit
@@ -1,3 +1,4 @@
+cd "$(git rev-parse --show-toplevel)/app"
 npm run lint
 npm run test-unit
 npm run test-storybook
