+++ app/src/features/card/components/Card.tsx
@@ -1,28 +1,8 @@
-import { createContext, useState } from "react";
+import { useState } from "react";
 import { CardPreview } from "./preview/CardPreview.tsx";
 import { CardForm } from "./form/CardForm.tsx";
+import { CardContext } from "./CardContext";
 import styled from "@emotion/styled";
-
-interface CardExpiryDateType {
-  "expiry-month": string;
-  "expiry-year": string;
-}
-
-interface CardNumberType {
-  "first-digits": string;
-  "second-digits": string;
-  "third-digits": string;
-  "fourth-digits": string;
-}
-
-interface CardContextType {
-  cardNumber: CardNumberType;
-  cardExpiryDate: CardExpiryDateType;
-  setCardNumber: React.Dispatch<React.SetStateAction<CardNumberType>>;
-  setCardExpiryDate: React.Dispatch<React.SetStateAction<CardExpiryDateType>>;
-}
-
-export const CardContext = createContext<CardContextType>(null!);
 
 export function Card() {
   const [cardNumber, setCardNumber] = useState({
