+++ app/src/features/card/components/form/CardExpiryDateInput.tsx
@@ -3,7 +3,7 @@ import { CardInput } from "./CardInput";
 import { Validator } from "../../validators/CardValidator";
 import { sanitizeErrors } from "../../../../Utils";
 import { CARD_INPUT } from "../../Constants";
-import { CardContext } from "../Card";
+import { CardContext } from "../CardContext";
 import { useState, useContext } from "react";
 import { ErrorMessage } from "./ErrorMessage";
