+++ app/src/features/card/components/preview/CardNetworkBrand.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { CardContext } from "../Card";
+import { CardContext } from "../CardContext";
 
 import { CardNetworkBrand } from "./CardNetworkBrand";
