+++ app/src/features/card/components/preview/CardNumber.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import { useContext } from "react";
-import { CardContext } from "../Card";
+import { CardContext } from "../CardContext";
 
 export function CardNumber() {
   const { cardNumber } = useContext(CardContext);
